feature

+++ src/main/java/gov/ao/usp/features/solicitacao/modelo/dto/SolicitacaoEstatisticasResponse.java
@@ -2,13 +2,12 @@ package gov.ao.usp.features.solicitacao.modelo.dto;
 
 public record SolicitacaoEstatisticasResponse(
 
-    long total,
-    long solicitacao,
-    long aprovadas,
-    long rejeitadas,
-    long devolvidas,
-    long visualizadas,
-    long canceladas
-
+    Integer total,
+    Integer solicitacao,
+    Integer aprovadas,
+    Integer rejeitadas,
+    Integer devolvidas,
+    Integer visualizadas,
+    Integer canceladas
 ) {
 }
